# internal/server/oapi/swagger.yaml
components:
  schemas:
    down.RemoveResult:
      properties:
        deleted:
          description: content IDs whose bytes were actually deleted
          items:
            type: string
          type: array
          uniqueItems: false
        untagged:
          description: tag refs removed; disk freed only when the last tag/content
            GCs
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
...
    retention.ApplyResult:
      properties:
        deleted:
          description: content-hash IDs whose bytes were freed
          items:
            type: string
          type: array
          uniqueItems: false
        errors:
          description: 'per-ref removal failures, "<ref>: <err>"'
          items:
            type: string
          type: array
          uniqueItems: false
        evaluated:
          description: number of records considered (delete+keep)
          type: integer
        untagged:
          description: refs whose tag was removed but content may remain
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
    retention.Candidate:
      properties:
        digest:
          type: string
        last_used:
          type: string
        reason:
          description: age_exceeded
          type: string
        ref:
          type: string
      type: object
    retention.Decision:
      properties:
        delete:
          items:
            $ref: '#/components/schemas/retention.Candidate'
          type: array
          uniqueItems: false
        keep:
          items:
            $ref: '#/components/schemas/retention.Kept'
          type: array
          uniqueItems: false
      type: object
    retention.Kept:
      properties:
        reason:
          description: in_use | pinned | keep_n_recent | within_max_age | grace |
            age_gc_disabled
          type: string
        ref:
          type: string
      type: object
...
    server.errorResponse:
      properties:
        error:
          type: string
      type: object
    server.gcRequest:
      properties:
        keep_n:
          description: Override how many most-recent tags to keep per repo; 0 disables
            keep-N.
          example: 3
          type: integer
        max_age:
          description: Override max image age (Go duration, e.g. "720h"); "0s" disables
            age GC.
          example: 720h
          type: string
        pins:
          description: Override the pinned references exempt from GC.
          example:
          - docker.io/library/nginx:1.27
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
...
    server.pinListResponse:
      properties:
        pins:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
    server.pinRequest:
      properties:
        ref:
          description: Image reference to pin or unpin (exact-match, required).
          example: docker.io/library/nginx:1.27
          type: string
      required:
      - ref
      type: object
    server.removeRequest:
      properties:
        ref:
          description: Image reference to delete from the engine store (required).
          example: docker.io/library/nginx:1.27
          type: string
      required:
      - ref
      type: object
    server.storeListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/store.Status'
          type: array
          uniqueItems: false
      type: object
    server.storePullRequest:
      properties:
        ref:
          description: Image reference the engine store should pull (required).
          example: docker.io/library/nginx:1.27
          type: string
      required:
      - ref
      type: object
    server.storePullResponse:
      properties:
        kind:
          type: string
        ref:
          type: string
        state:
          type: string
        store:
          type: string
      type: object
    store.Caps:
      description: what this store can do
      properties:
        gc:
          description: engine supports image GC (phase 2)
          type: boolean
        pull:
          description: engine can be triggered to pull
          type: boolean
        read:
          description: 'registry: pull blobs'
          type: boolean
        verify:
          description: engine can verify signatures (phase 2)
          type: boolean
        write:
          description: 'registry: push blobs'
          type: boolean
      type: object
    store.Status:
      properties:
        address:
          type: string
        capabilities:
          $ref: '#/components/schemas/store.Caps'
        error:
          description: engine readiness error, if not ready
          type: string
        host:
          type: string
        kind:
          enum:
          - oci
          - docker
          - containerd
          type: string
        mode:
          type: string
        name:
          type: string
        namespace:
          type: string
        ready:
          description: 'registries: always true (from config); engines: live Ready()
            probe'
          type: boolean
      type: object
...
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: |-
    Move container images between stores (OCI registries and docker/containerd engines) and track per-layer progress.
    A job copies an image `from` an OCI store `to` another, then `distribute` engines pull it.
  title: gantry API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /v1/store:
    get:
      description: Configured stores with their kind, capabilities, and readiness.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.storeListResponse'
          description: OK
      security:
      - BearerAuth: []
      summary: List stores
      tags:
      - stores
  /v1/store/{name}/gc:
    get:
      description: 'Returns the retention decision (keep/delete) without deleting
        anything. An optional body overrides the configured max_age/keep_n/pins for
        this call. NOTE: a GET request body is dropped by fetch()/XHR, some HTTP clients,
        and proxies — to pass overrides reliably, use POST.'
      parameters:
      - description: engine store name
        in: path
        name: name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/server.gcRequest'
        description: policy overrides
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/retention.Decision'
          description: dry-run keep/delete decision
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Not Found
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Not Implemented
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Bad Gateway
      security:
      - BearerAuth: []
      summary: Evaluate retention GC for a store (dry-run)
      tags:
      - retention
    post:
      description: Applies the deletions and returns the apply result. An optional
        body overrides the configured max_age/keep_n/pins for this call.
      parameters:
      - description: engine store name
        in: path
        name: name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/server.gcRequest'
        description: policy overrides
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/retention.ApplyResult'
          description: deletions applied
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Not Found
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Not Implemented
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Bad Gateway
      security:
      - BearerAuth: []
      summary: Run retention GC for a store (apply)
      tags:
      - retention
...
  /v1/store/{name}/pin:
    delete:
      parameters:
      - description: engine store name
        in: path
        name: name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/server.pinRequest'
        description: reference to unpin
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Not Implemented
      security:
      - BearerAuth: []
      summary: Unpin a reference
      tags:
      - retention
    get:
      description: Pinned references are exempt from retention GC (exact-match).
      parameters:
      - description: engine store name
        in: path
        name: name
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.pinListResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Not Implemented
      security:
      - BearerAuth: []
      summary: List pinned references for a store
      tags:
      - retention
    post:
      parameters:
      - description: engine store name
        in: path
        name: name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/server.pinRequest'
        description: reference to pin
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Not Implemented
      security:
      - BearerAuth: []
      summary: Pin a reference (exempt from GC)
      tags:
      - retention
  /v1/store/{name}/pull:
    post:
      description: Tells one engine store to pull a reference, decoupled from the
        job pipeline (manual reconcile).
      parameters:
      - description: engine store name
        in: path
        name: name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/server.storePullRequest'
        description: reference to pull
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.storePullResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Not Found
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Bad Gateway
      security:
      - BearerAuth: []
      summary: Trigger an engine store to pull
      tags:
      - stores
  /v1/store/{name}/remove:
    post:
      description: Manually deletes one image from an engine store and syncs the retention
        index.
      parameters:
      - description: engine store name
        in: path
        name: name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/server.removeRequest'
        description: reference to remove
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/down.RemoveResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Not Found
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/server.errorResponse'
          description: Bad Gateway
      security:
      - BearerAuth: []
      summary: Remove an image from an engine store
      tags:
      - retention